♻️ refactor(auth-files): move auth files pages into feature module
Move auth file route pages, OAuth edit screens, page styles, and page tests into src/features/authFiles.

This keeps route files as thin exports and removes feature component imports from page-owned styles.
