+++ src/features/authFiles/AuthFilesOAuthExcludedEditPage.module.scss
@@ -1,5 +1,5 @@
-@use '../styles/variables' as *;
-@use '../styles/mixins' as *;
+@use '../../styles/variables' as *;
+@use '../../styles/mixins' as *;
 
 .pageContent {
   width: 100%;
@@ -141,41 +141,26 @@
   }
 }
 
-.mappingsHeader {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  gap: $spacing-md;
-  padding: $spacing-md $spacing-lg;
-  border-bottom: 1px solid var(--border-color);
-
-  @include mobile {
-    padding-left: $spacing-md;
-    padding-right: $spacing-md;
-  }
-}
-
-.mappingsTitle {
-  font-weight: 700;
-  color: var(--text-primary);
-}
-
 .modelsHint {
   display: flex;
   align-items: center;
   gap: $spacing-xs;
-  padding: $spacing-sm $spacing-lg;
   font-size: 13px;
   color: var(--text-secondary);
-  border-bottom: 1px solid var(--border-color);
-
-  @include mobile {
-    padding-left: $spacing-md;
-    padding-right: $spacing-md;
-  }
 }
 
-.mappingsBody {
+.loadingModels {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  gap: $spacing-sm;
+  padding: $spacing-xl 0;
+  color: var(--text-secondary);
+}
+
+.modelList {
+  max-height: 520px;
+  overflow: auto;
   padding: $spacing-sm $spacing-lg $spacing-lg;
 
   @include mobile {
@@ -184,42 +169,52 @@
   }
 }
 
-.mappingRow {
-  display: grid;
-  grid-template-columns: 1fr auto 1fr auto auto;
-  align-items: center;
-  gap: $spacing-sm;
+.modelItem {
+  width: 100%;
+  align-items: flex-start;
   padding: 10px 0;
   border-bottom: 1px solid var(--border-color);
+  border-radius: $radius-sm;
+  transition: background-color $transition-fast;
 
   &:last-child {
     border-bottom: none;
   }
 
-  @include mobile {
-    grid-template-columns: 1fr;
-    gap: $spacing-sm;
+  &:hover {
+    background-color: var(--bg-hover);
   }
 }
 
-.mappingSeparator {
+.modelText {
+  display: flex;
+  flex-direction: column;
+  gap: 2px;
+  min-width: 0;
+  flex: 1;
+}
+
+.modelId {
+  font-size: 13px;
+  font-weight: 600;
+  color: var(--text-primary);
+  word-break: break-all;
+}
+
+.modelDisplayName {
+  font-size: 12px;
   color: var(--text-secondary);
+  word-break: break-all;
+}
+
+.emptyModels {
+  padding: $spacing-xl $spacing-lg;
+  color: var(--text-secondary);
+  font-size: 13px;
   text-align: center;
 
   @include mobile {
-    display: none;
-  }
-}
-
-.mappingAliasInput {
-  width: 100%;
-}
-
-.mappingFork {
-  display: flex;
-  align-items: center;
-
-  @include mobile {
-    justify-content: flex-start;
+    padding-left: $spacing-md;
+    padding-right: $spacing-md;
   }
 }
