+++ src/features/authFiles/AuthFilesPage.module.scss
@@ -1,5 +1,5 @@
-@use '../styles/variables' as *;
-@use '../styles/mixins' as *;
+@use '../../styles/variables' as *;
+@use '../../styles/mixins' as *;
 
 .container {
   display: flex;
