+++ src/features/authFiles/AuthFilesPage.tsx
@@ -1,0 +1,1215 @@
+import {
+  useCallback,
+  type CSSProperties,
+  useEffect,
+  useLayoutEffect,
+  useMemo,
+  useRef,
+  useState,
+  type ChangeEvent,
+} from 'react';
+import { createPortal } from 'react-dom';
+import { useTranslation } from 'react-i18next';
+import type { TFunction } from 'i18next';
+import { useNavigate } from 'react-router-dom';
+import { animate } from 'motion/mini';
+import type { AnimationPlaybackControlsWithThen } from 'motion-dom';
+import type { AuthFileItem, CodexQuotaState } from '@/types';
+import { useInterval } from '@/hooks/useInterval';
+import { useHeaderRefresh } from '@/hooks/useHeaderRefresh';
+import { usePageTransitionLayer } from '@/components/common/PageTransitionLayer';
+import { Button } from '@/components/ui/Button';
+import { Input } from '@/components/ui/Input';
+import { Select } from '@/components/ui/Select';
+import { IconFilterAll, IconSearch } from '@/components/ui/icons';
+import { EmptyState } from '@/components/ui/EmptyState';
+import { ToggleSwitch } from '@/components/ui/ToggleSwitch';
+import { copyToClipboard } from '@/utils/clipboard';
+import {
+  normalizePlanType,
+  resolveCodexChatgptAccountId,
+  resolveCodexPlanType,
+} from '@/utils/quota';
+import {
+  MAX_CARD_PAGE_SIZE,
+  MIN_CARD_PAGE_SIZE,
+  QUOTA_PROVIDER_TYPES,
+  clampCardPageSize,
+  getAuthFileIcon,
+  getTypeColor,
+  getTypeLabel,
+  hasAuthFileStatusMessage,
+  isHealthyAuthFile,
+  isRuntimeOnlyAuthFile,
+  normalizeProviderKey,
+  parsePriorityValue,
+  type QuotaProviderType,
+  type ResolvedTheme,
+} from '@/features/authFiles/constants';
+import { AuthFileCard } from '@/features/authFiles/components/AuthFileCard';
+import { AuthJsonPasteModal } from '@/features/authFiles/components/AuthJsonPasteModal';
+import { AuthFileModelsModal } from '@/features/authFiles/components/AuthFileModelsModal';
+import { AuthFilesPrefixProxyEditorModal } from '@/features/authFiles/components/AuthFilesPrefixProxyEditorModal';
+import { OAuthExcludedCard } from '@/features/authFiles/components/OAuthExcludedCard';
+import { OAuthModelAliasCard } from '@/features/authFiles/components/OAuthModelAliasCard';
+import { useAuthFilesData } from '@/features/authFiles/hooks/useAuthFilesData';
+import { useAuthFilesModels } from '@/features/authFiles/hooks/useAuthFilesModels';
+import { useAuthFilesOauth } from '@/features/authFiles/hooks/useAuthFilesOauth';
+import { useAuthFilesPrefixProxyEditor } from '@/features/authFiles/hooks/useAuthFilesPrefixProxyEditor';
+import { useAuthFilesStatusBarCache } from '@/features/authFiles/hooks/useAuthFilesStatusBarCache';
+import {
+  normalizeAuthFilesSortMode,
+  readAuthFilesUiState,
+  readPersistedAuthFilesCompactMode,
+  writeAuthFilesUiState,
+  writePersistedAuthFilesCompactMode,
+  type AuthFilesSortMode,
+} from '@/features/authFiles/uiState';
+import type { AuthJsonInputType } from '@/features/authFiles/sessionAuthConverter';
+import { useAuthStore, useNotificationStore, useQuotaStore, useThemeStore } from '@/stores';
+import styles from './AuthFilesPage.module.scss';
+
+const easePower3Out = (progress: number) => 1 - (1 - progress) ** 4;
+const easePower2In = (progress: number) => progress ** 3;
+const BATCH_BAR_BASE_TRANSFORM = 'translateX(-50%)';
+const BATCH_BAR_HIDDEN_TRANSFORM = 'translateX(-50%) translateY(56px)';
+const DEFAULT_REGULAR_PAGE_SIZE = 9;
+const DEFAULT_COMPACT_PAGE_SIZE = 12;
+
+const escapeWildcardSearchSegment = (value: string) => value.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
+
+const buildWildcardSearch = (value: string): RegExp | null => {
+  if (!value.includes('*')) return null;
+  const pattern = value.split('*').map(escapeWildcardSearchSegment).join('.*');
+  return new RegExp(pattern, 'i');
+};
+
+const PREMIUM_CODEX_PLAN_TYPES = new Set(['pro', 'prolite', 'pro-lite', 'pro_lite']);
+
+const compareAuthFileName = (left: { name: string }, right: { name: string }) =>
+  left.name.localeCompare(right.name, undefined, { numeric: true, sensitivity: 'base' });
+
+const getAuthFileNoteValue = (file: AuthFileItem): string => {
+  const raw = file.note ?? file['note'];
+  if (raw === undefined || raw === null) return '';
+  return String(raw).trim();
+};
+
+const compareAuthFileNote = (
+  left: AuthFileItem,
+  right: AuthFileItem,
+  direction: 'asc' | 'desc'
+) => {
+  const leftNote = getAuthFileNoteValue(left);
+  const rightNote = getAuthFileNoteValue(right);
+  const leftKnown = leftNote.length > 0;
+  const rightKnown = rightNote.length > 0;
+
+  if (leftKnown || rightKnown) {
+    if (!leftKnown) return 1;
+    if (!rightKnown) return -1;
+    const diff = leftNote.localeCompare(rightNote, undefined, {
+      numeric: true,
+      sensitivity: 'base',
+    });
+    if (diff !== 0) return direction === 'asc' ? diff : -diff;
+  }
+
+  return compareAuthFileName(left, right);
+};
+
+const compareAuthFilePriority = (
+  left: AuthFileItem,
+  right: AuthFileItem,
+  direction: 'asc' | 'desc'
+) => {
+  const leftPriority = parsePriorityValue(left.priority ?? left['priority']);
+  const rightPriority = parsePriorityValue(right.priority ?? right['priority']);
+  const leftKnown = leftPriority !== undefined;
+  const rightKnown = rightPriority !== undefined;
+
+  if (leftKnown || rightKnown) {
+    if (!leftKnown) return 1;
+    if (!rightKnown) return -1;
+    const leftValue = leftPriority ?? 0;
+    const rightValue = rightPriority ?? 0;
+    const diff =
+      direction === 'desc'
+        ? rightValue - leftValue
+        : leftValue - rightValue;
+    if (diff !== 0) return diff;
+  }
+
+  return compareAuthFileName(left, right);
+};
+
+const stringifySearchValue = (value: unknown): string[] => {
+  if (value === undefined || value === null) return [];
+  if (Array.isArray(value)) return value.flatMap(stringifySearchValue);
+  if (typeof value === 'string') return value.trim() ? [value] : [];
+  if (typeof value === 'number' || typeof value === 'boolean') return [String(value)];
+  return [];
+};
+
+const getCodexPlanLabel = (planType: string | null | undefined, t: TFunction): string | null => {
+  const normalized = normalizePlanType(planType);
+  if (!normalized) return null;
+  if (normalized === 'pro') return t('codex_quota.plan_pro');
+  if (PREMIUM_CODEX_PLAN_TYPES.has(normalized) && normalized !== 'pro') {
+    return t('codex_quota.plan_prolite');
+  }
+  if (normalized === 'plus') return t('codex_quota.plan_plus');
+  if (normalized === 'team') return t('codex_quota.plan_team');
+  if (normalized === 'free') return t('codex_quota.plan_free');
+  return planType || normalized;
+};
+
+const getAuthFilePlanType = (file: AuthFileItem, quota?: CodexQuotaState): string | null =>
+  resolveCodexPlanType(file) ?? quota?.planType ?? null;
+
+const getAuthFilePlanSortRank = (file: AuthFileItem, quota?: CodexQuotaState): number | null => {
+  const normalized = normalizePlanType(getAuthFilePlanType(file, quota));
+  if (!normalized) return null;
+  if (normalized === 'pro') return 50;
+  if (PREMIUM_CODEX_PLAN_TYPES.has(normalized) && normalized !== 'pro') return 40;
+  if (normalized === 'team') return 30;
+  if (normalized === 'plus') return 20;
+  if (normalized === 'free') return 10;
+  return 0;
+};
+
+const getAuthFileSearchValues = (file: AuthFileItem, t: TFunction, quota?: CodexQuotaState) => {
+  const planType = getAuthFilePlanType(file, quota);
+  const planLabel = getCodexPlanLabel(planType, t);
+  const accountId = resolveCodexChatgptAccountId(file);
+  const type = file.type || file.provider;
+
+  return [
+    file.name,
+    file.type,
+    file.provider,
+    type ? getTypeLabel(t, String(type)) : null,
+    file.authIndex,
+    file['auth_index'],
+    file.status,
+    file.state,
+    file.statusMessage,
+    file['status_message'],
+    file.error,
+    file.errorStatus,
+    file['error_status'],
+    quota?.status,
+    quota?.error,
+    quota?.errorStatus,
+    planType,
+    planLabel,
+    accountId,
+  ];
+};
+
+export function AuthFilesPage() {
+  const { t } = useTranslation();
+  const showNotification = useNotificationStore((state) => state.showNotification);
+  const connectionStatus = useAuthStore((state) => state.connectionStatus);
+  const resolvedTheme: ResolvedTheme = useThemeStore((state) => state.resolvedTheme);
+  const codexQuota = useQuotaStore((state) => state.codexQuota);
+  const pageTransitionLayer = usePageTransitionLayer();
+  const isCurrentLayer = pageTransitionLayer ? pageTransitionLayer.status === 'current' : true;
+  const navigate = useNavigate();
+
+  const [filter, setFilter] = useState<'all' | string>('all');
+  const [problemOnly, setProblemOnly] = useState(false);
+  const [disabledOnly, setDisabledOnly] = useState(false);
+  const [healthyOnly, setHealthyOnly] = useState(false);
+  const [compactMode, setCompactMode] = useState(false);
+  const [search, setSearch] = useState('');
+  const [page, setPage] = useState(1);
+  const [pageSizeByMode, setPageSizeByMode] = useState({
+    regular: DEFAULT_REGULAR_PAGE_SIZE,
+    compact: DEFAULT_COMPACT_PAGE_SIZE,
+  });
+  const [pageSizeInput, setPageSizeInput] = useState('9');
+  const [viewMode, setViewMode] = useState<'diagram' | 'list'>('list');
+  const [sortMode, setSortMode] = useState<AuthFilesSortMode>('default');
+  const [batchActionBarVisible, setBatchActionBarVisible] = useState(false);
+  const [uiStateHydrated, setUiStateHydrated] = useState(false);
+  const [authJsonPasteOpen, setAuthJsonPasteOpen] = useState(false);
+  const floatingBatchActionsRef = useRef<HTMLDivElement>(null);
+  const batchActionAnimationRef = useRef<AnimationPlaybackControlsWithThen | null>(null);
+  const previousSelectionCountRef = useRef(0);
+  const selectionCountRef = useRef(0);
+
+  const {
+    files,
+    selectedFiles,
+    selectionCount,
+    loading,
+    error,
+    uploading,
+    authJsonPasteSaving,
+    deleting,
+    deletingAll,
+    statusUpdating,
+    batchStatusUpdating,
+    fileInputRef,
+    loadFiles,
+    handleUploadClick,
+    handleFileChange,
+    savePastedAuthJson,
+    handleDelete,
+    handleDeleteAll,
+    handleDownload,
+    handleStatusToggle,
+    toggleSelect,
+    selectAllVisible,
+    invertVisibleSelection,
+    deselectAll,
+    batchDownload,
+    batchSetStatus,
+    batchDelete,
+  } = useAuthFilesData();
+
+  const statusBarCache = useAuthFilesStatusBarCache(files);
+
+  const {
+    excluded,
+    excludedError,
+    modelAlias,
+    modelAliasError,
+    allProviderModels,
+    loadExcluded,
+    loadModelAlias,
+    deleteExcluded,
+    deleteModelAlias,
+    handleMappingUpdate,
+    handleDeleteLink,
+    handleToggleFork,
+    handleRenameAlias,
+    handleDeleteAlias,
+  } = useAuthFilesOauth({ viewMode, files });
+
+  const {
+    modelsModalOpen,
+    modelsLoading,
+    modelsList,
+    modelsFileName,
+    modelsFileType,
+    modelsError,
+    showModels,
+    closeModelsModal,
+  } = useAuthFilesModels();
+
+  const {
+    prefixProxyEditor,
+    prefixProxyUpdatedText,
+    prefixProxyDirty,
+    openPrefixProxyEditor,
+    closePrefixProxyEditor,
+    handlePrefixProxyChange,
+    handlePrefixProxySave,
+  } = useAuthFilesPrefixProxyEditor({
+    disableControls: connectionStatus !== 'connected',
+    loadFiles,
+  });
+
+  const disableControls = connectionStatus !== 'connected';
+  const normalizedFilter = normalizeProviderKey(String(filter));
+  const quotaFilterType: QuotaProviderType | null = QUOTA_PROVIDER_TYPES.has(
+    normalizedFilter as QuotaProviderType
+  )
+    ? (normalizedFilter as QuotaProviderType)
+    : null;
+  const pageSize = compactMode ? pageSizeByMode.compact : pageSizeByMode.regular;
+
+  useEffect(() => {
+    const persistedCompactMode = readPersistedAuthFilesCompactMode();
+    if (typeof persistedCompactMode === 'boolean') {
+      setCompactMode(persistedCompactMode);
+    }
+
+    const persisted = readAuthFilesUiState();
+    if (persisted) {
+      if (typeof persisted.filter === 'string' && persisted.filter.trim()) {
+        setFilter(normalizeProviderKey(persisted.filter));
+      }
+      if (typeof persisted.problemOnly === 'boolean') {
+        setProblemOnly(persisted.problemOnly);
+      }
+      if (typeof persisted.disabledOnly === 'boolean') {
+        setDisabledOnly(persisted.disabledOnly);
+      }
+      if (typeof persisted.healthyOnly === 'boolean') {
+        setHealthyOnly(persisted.healthyOnly);
+      }
+      if (
+        typeof persistedCompactMode !== 'boolean' &&
+        typeof persisted.compactMode === 'boolean'
+      ) {
+        setCompactMode(persisted.compactMode);
+      }
+      if (typeof persisted.search === 'string') {
+        setSearch(persisted.search);
+      }
+      if (typeof persisted.page === 'number' && Number.isFinite(persisted.page)) {
+        setPage(Math.max(1, Math.round(persisted.page)));
+      }
+      const legacyPageSize =
+        typeof persisted.pageSize === 'number' && Number.isFinite(persisted.pageSize)
+          ? clampCardPageSize(persisted.pageSize)
+          : null;
+      const regularPageSize =
+        typeof persisted.regularPageSize === 'number' && Number.isFinite(persisted.regularPageSize)
+          ? clampCardPageSize(persisted.regularPageSize)
+          : (legacyPageSize ?? DEFAULT_REGULAR_PAGE_SIZE);
+      const compactPageSize =
+        typeof persisted.compactPageSize === 'number' && Number.isFinite(persisted.compactPageSize)
+          ? clampCardPageSize(persisted.compactPageSize)
+          : (legacyPageSize ?? DEFAULT_COMPACT_PAGE_SIZE);
+      setPageSizeByMode({
+        regular: regularPageSize,
+        compact: compactPageSize,
+      });
+      const persistedSortMode = normalizeAuthFilesSortMode(persisted.sortMode);
+      if (persistedSortMode) {
+        setSortMode(persistedSortMode);
+      }
+    }
+
+    setUiStateHydrated(true);
+  }, []);
+
+  useEffect(() => {
+    if (!uiStateHydrated) return;
+
+    writeAuthFilesUiState({
+      filter,
+      problemOnly,
+      disabledOnly,
+      healthyOnly,
+      compactMode,
+      search,
+      page,
+      pageSize,
+      regularPageSize: pageSizeByMode.regular,
+      compactPageSize: pageSizeByMode.compact,
+      sortMode,
+    });
+    writePersistedAuthFilesCompactMode(compactMode);
+  }, [
+    compactMode,
+    disabledOnly,
+    filter,
+    healthyOnly,
+    page,
+    pageSize,
+    pageSizeByMode,
+    problemOnly,
+    search,
+    sortMode,
+    uiStateHydrated,
+  ]);
+
+  useEffect(() => {
+    setPageSizeInput(String(pageSize));
+  }, [pageSize]);
+
+  const setCurrentModePageSize = useCallback(
+    (next: number) => {
+      setPageSizeByMode((current) =>
+        compactMode ? { ...current, compact: next } : { ...current, regular: next }
+      );
+    },
+    [compactMode]
+  );
+
+  const commitPageSizeInput = (rawValue: string) => {
+    const trimmed = rawValue.trim();
+    if (!trimmed) {
+      setPageSizeInput(String(pageSize));
+      return;
+    }
+
+    const value = Number(trimmed);
+    if (!Number.isFinite(value)) {
+      setPageSizeInput(String(pageSize));
+      return;
+    }
+
+    const next = clampCardPageSize(value);
+    setCurrentModePageSize(next);
+    setPageSizeInput(String(next));
+    setPage(1);
+  };
+
+  const handlePageSizeChange = (event: ChangeEvent<HTMLInputElement>) => {
+    const rawValue = event.currentTarget.value;
+    setPageSizeInput(rawValue);
+
+    const trimmed = rawValue.trim();
+    if (!trimmed) return;
+
+    const parsed = Number(trimmed);
+    if (!Number.isFinite(parsed)) return;
+
+    const rounded = Math.round(parsed);
+    if (rounded < MIN_CARD_PAGE_SIZE || rounded > MAX_CARD_PAGE_SIZE) return;
+
+    setCurrentModePageSize(rounded);
+    setPage(1);
+  };
+
+  const handleSortModeChange = useCallback(
+    (value: string) => {
+      const nextSortMode = normalizeAuthFilesSortMode(value);
+      if (!nextSortMode || nextSortMode === sortMode) return;
+      setSortMode(nextSortMode);
+      setPage(1);
+      void loadFiles().catch(() => {});
+    },
+    [loadFiles, sortMode]
+  );
+
+  const handleSavePastedAuthJson = useCallback(
+    async (type: AuthJsonInputType, fileName: string, jsonText: string) => {
+      await savePastedAuthJson(type, fileName, jsonText);
+      setAuthJsonPasteOpen(false);
+    },
+    [savePastedAuthJson]
+  );
+
+  const handleHeaderRefresh = useCallback(async () => {
+    await Promise.all([loadFiles(), loadExcluded(), loadModelAlias()]);
+  }, [loadFiles, loadExcluded, loadModelAlias]);
+
+  useHeaderRefresh(handleHeaderRefresh);
+
+  useEffect(() => {
+    if (!isCurrentLayer) return;
+    loadFiles();
+    loadExcluded();
+    loadModelAlias();
+  }, [isCurrentLayer, loadFiles, loadExcluded, loadModelAlias]);
+
+  useInterval(
+    () => {
+      void loadFiles().catch(() => {});
+    },
+    isCurrentLayer ? 240_000 : null
+  );
+
+  const existingTypes = useMemo(() => {
+    const types = new Set<string>(['all']);
+    files.forEach((file) => {
+      const type = normalizeProviderKey(String(file.type ?? file.provider ?? ''));
+      if (type) types.add(type);
+    });
+    return Array.from(types);
+  }, [files]);
+
+  const filesMatchingStatusFilters = useMemo(
+    () =>
+      files.filter((file) => {
+        if (problemOnly && !hasAuthFileStatusMessage(file)) return false;
+        if (disabledOnly && file.disabled !== true) return false;
+        if (healthyOnly && !isHealthyAuthFile(file)) return false;
+        return true;
+      }),
+    [disabledOnly, files, healthyOnly, problemOnly]
+  );
+
+  const sortOptions = useMemo(
+    () => [
+      { value: 'default', label: t('auth_files.sort_default') },
+      { value: 'name-asc', label: t('auth_files.sort_name_asc') },
+      { value: 'note-asc', label: t('auth_files.sort_note_asc') },
+      { value: 'note-desc', label: t('auth_files.sort_note_desc') },
+      { value: 'priority-desc', label: t('auth_files.sort_priority_desc') },
+      { value: 'priority-asc', label: t('auth_files.sort_priority_asc') },
+      { value: 'plan-desc', label: t('auth_files.sort_plan_desc') },
+      { value: 'plan-asc', label: t('auth_files.sort_plan_asc') },
+    ],
+    [t]
+  );
+
+  const typeCounts = useMemo(() => {
+    const counts: Record<string, number> = { all: filesMatchingStatusFilters.length };
+    filesMatchingStatusFilters.forEach((file) => {
+      const type = normalizeProviderKey(String(file.type ?? file.provider ?? ''));
+      if (!type) return;
+      counts[type] = (counts[type] || 0) + 1;
+    });
+    return counts;
+  }, [filesMatchingStatusFilters]);
+
+  const normalizedSearch = search.trim();
+  const wildcardSearch = useMemo(() => buildWildcardSearch(normalizedSearch), [normalizedSearch]);
+
+  const filtered = useMemo(() => {
+    const normalizedTerm = normalizedSearch.toLowerCase();
+
+    return filesMatchingStatusFilters.filter((item) => {
+      const type = normalizeProviderKey(String(item.type ?? item.provider ?? ''));
+      const matchType = normalizedFilter === 'all' || type === normalizedFilter;
+      const matchSearch =
+        !normalizedSearch ||
+        stringifySearchValue(getAuthFileSearchValues(item, t, codexQuota[item.name])).some(
+          (value) => {
+            const content = value.toString();
+            return wildcardSearch
+              ? wildcardSearch.test(content)
+              : content.toLowerCase().includes(normalizedTerm);
+          }
+        );
+      return matchType && matchSearch;
+    });
+  }, [codexQuota, filesMatchingStatusFilters, normalizedFilter, normalizedSearch, t, wildcardSearch]);
+
+  const sorted = useMemo(() => {
+    const copy = [...filtered];
+    if (sortMode === 'default') {
+      copy.sort((a, b) => {
+        const providerA = normalizeProviderKey(String(a.provider ?? a.type ?? 'unknown'));
+        const providerB = normalizeProviderKey(String(b.provider ?? b.type ?? 'unknown'));
+        const providerCompare = providerA.localeCompare(providerB);
+        if (providerCompare !== 0) return providerCompare;
+        return compareAuthFileName(a, b);
+      });
+    } else if (sortMode === 'name-asc') {
+      copy.sort(compareAuthFileName);
+    } else if (sortMode === 'note-asc' || sortMode === 'note-desc') {
+      copy.sort((a, b) =>
+        compareAuthFileNote(a, b, sortMode === 'note-desc' ? 'desc' : 'asc')
+      );
+    } else if (sortMode === 'priority-asc' || sortMode === 'priority-desc') {
+      copy.sort((a, b) =>
+        compareAuthFilePriority(a, b, sortMode === 'priority-desc' ? 'desc' : 'asc')
+      );
+    } else if (sortMode === 'plan-asc' || sortMode === 'plan-desc') {
+      copy.sort((a, b) => {
+        const leftRank = getAuthFilePlanSortRank(a, codexQuota[a.name]);
+        const rightRank = getAuthFilePlanSortRank(b, codexQuota[b.name]);
+        const leftKnown = leftRank !== null && leftRank !== undefined;
+        const rightKnown = rightRank !== null && rightRank !== undefined;
+
+        if (leftKnown || rightKnown) {
+          if (!leftKnown) return 1;
+          if (!rightKnown) return -1;
+          const rankDiff = sortMode === 'plan-desc' ? rightRank - leftRank : leftRank - rightRank;
+          if (rankDiff !== 0) return rankDiff;
+        }
+
+        return compareAuthFileName(a, b);
+      });
+    }
+    return copy;
+  }, [codexQuota, filtered, sortMode]);
+
+  const totalPages = Math.max(1, Math.ceil(sorted.length / pageSize));
+  const currentPage = Math.min(page, totalPages);
+  const start = (currentPage - 1) * pageSize;
+  const pageItems = sorted.slice(start, start + pageSize);
+  const selectablePageItems = useMemo(
+    () => pageItems.filter((file) => !isRuntimeOnlyAuthFile(file)),
+    [pageItems]
+  );
+  const selectableFilteredItems = useMemo(
+    () => sorted.filter((file) => !isRuntimeOnlyAuthFile(file)),
+    [sorted]
+  );
+  const selectedNames = useMemo(() => Array.from(selectedFiles), [selectedFiles]);
+  const selectedHasStatusUpdating = useMemo(
+    () => selectedNames.some((name) => statusUpdating[name] === true),
+    [selectedNames, statusUpdating]
+  );
+  const batchStatusButtonsDisabled =
+    disableControls ||
+    selectedNames.length === 0 ||
+    batchStatusUpdating ||
+    selectedHasStatusUpdating;
+
+  const copyTextWithNotification = useCallback(
+    async (text: string) => {
+      const copied = await copyToClipboard(text);
+      showNotification(
+        copied
+          ? t('notification.link_copied', { defaultValue: 'Copied to clipboard' })
+          : t('notification.copy_failed', { defaultValue: 'Copy failed' }),
+        copied ? 'success' : 'error'
+      );
+    },
+    [showNotification, t]
+  );
+
+  const openExcludedEditor = useCallback(
+    (provider?: string) => {
+      const providerValue = (provider || (filter !== 'all' ? String(filter) : '')).trim();
+      const params = new URLSearchParams();
+      if (providerValue) {
+        params.set('provider', providerValue);
+      }
+      const nextSearch = params.toString();
+      navigate(`/auth-files/oauth-excluded${nextSearch ? `?${nextSearch}` : ''}`, {
+        state: { fromAuthFiles: true },
+      });
+    },
+    [filter, navigate]
+  );
+
+  const openModelAliasEditor = useCallback(
+    (provider?: string) => {
+      const providerValue = (provider || (filter !== 'all' ? String(filter) : '')).trim();
+      const params = new URLSearchParams();
+      if (providerValue) {
+        params.set('provider', providerValue);
+      }
+      const nextSearch = params.toString();
+      navigate(`/auth-files/oauth-model-alias${nextSearch ? `?${nextSearch}` : ''}`, {
+        state: { fromAuthFiles: true },
+      });
+    },
+    [filter, navigate]
+  );
+
+  useLayoutEffect(() => {
+    if (typeof window === 'undefined') return;
+
+    const actionsEl = floatingBatchActionsRef.current;
+    if (!actionsEl) {
+      document.documentElement.style.removeProperty('--auth-files-action-bar-height');
+      return;
+    }
+
+    const updatePadding = () => {
+      const height = actionsEl.getBoundingClientRect().height;
+      document.documentElement.style.setProperty('--auth-files-action-bar-height', `${height}px`);
+    };
+
+    updatePadding();
+    window.addEventListener('resize', updatePadding);
+
+    const ro = typeof ResizeObserver === 'undefined' ? null : new ResizeObserver(updatePadding);
+    ro?.observe(actionsEl);
+
+    return () => {
+      ro?.disconnect();
+      window.removeEventListener('resize', updatePadding);
+      document.documentElement.style.removeProperty('--auth-files-action-bar-height');
+    };
+  }, [batchActionBarVisible, selectionCount]);
+
+  useEffect(() => {
+    selectionCountRef.current = selectionCount;
+    if (selectionCount > 0) {
+      setBatchActionBarVisible(true);
+    }
+  }, [selectionCount]);
+
+  useLayoutEffect(() => {
+    if (!batchActionBarVisible) return;
+    const currentCount = selectionCount;
+    const previousCount = previousSelectionCountRef.current;
+    const actionsEl = floatingBatchActionsRef.current;
+    if (!actionsEl) return;
+
+    batchActionAnimationRef.current?.stop();
+    batchActionAnimationRef.current = null;
+
+    if (currentCount > 0 && previousCount === 0) {
+      batchActionAnimationRef.current = animate(
+        actionsEl,
+        {
+          transform: [BATCH_BAR_HIDDEN_TRANSFORM, BATCH_BAR_BASE_TRANSFORM],
+          opacity: [0, 1],
+        },
+        {
+          duration: 0.28,
+          ease: easePower3Out,
+          onComplete: () => {
+            actionsEl.style.transform = BATCH_BAR_BASE_TRANSFORM;
+            actionsEl.style.opacity = '1';
+          },
+        }
+      );
+    } else if (currentCount === 0 && previousCount > 0) {
+      batchActionAnimationRef.current = animate(
+        actionsEl,
+        {
+          transform: [BATCH_BAR_BASE_TRANSFORM, BATCH_BAR_HIDDEN_TRANSFORM],
+          opacity: [1, 0],
+        },
+        {
+          duration: 0.22,
+          ease: easePower2In,
+          onComplete: () => {
+            if (selectionCountRef.current === 0) {
+              setBatchActionBarVisible(false);
+            }
+          },
+        }
+      );
+    }
+
+    previousSelectionCountRef.current = currentCount;
+  }, [batchActionBarVisible, selectionCount]);
+
+  useEffect(
+    () => () => {
+      batchActionAnimationRef.current?.stop();
+      batchActionAnimationRef.current = null;
+    },
+    []
+  );
+
+  const renderFilterTags = () => (
+    <div className={styles.filterTags}>
+      {existingTypes.map((type) => {
+        const isActive = normalizedFilter === type;
+        const iconSrc = getAuthFileIcon(type, resolvedTheme);
+        const color =
+          type === 'all'
+            ? { bg: 'var(--color-primary-light-9)', text: 'var(--primary-color)' }
+            : getTypeColor(type, resolvedTheme);
+        const buttonStyle = {
+          '--filter-color': color.text,
+          '--filter-surface': color.bg,
+          '--filter-active-text': resolvedTheme === 'dark' ? '#111827' : '#ffffff',
+        } as CSSProperties;
+
+        return (
+          <button
+            key={type}
+            className={`${styles.filterTag} ${isActive ? styles.filterTagActive : ''}`}
+            style={buttonStyle}
+            onClick={() => {
+              setFilter(type);
+              setPage(1);
+            }}
+          >
+            <span className={styles.filterTagLabel}>
+              {type === 'all' ? (
+                <span className={`${styles.filterTagIconWrap} ${styles.filterAllIconWrap}`}>
+                  <IconFilterAll className={styles.filterAllIcon} size={16} />
+                </span>
+              ) : (
+                <span className={styles.filterTagIconWrap}>
+                  {iconSrc ? (
+                    <img src={iconSrc} alt="" className={styles.filterTagIcon} />
+                  ) : (
+                    <span className={styles.filterTagIconFallback}>
+                      {getTypeLabel(t, type).slice(0, 1).toUpperCase()}
+                    </span>
+                  )}
+                </span>
+              )}
+              <span className={styles.filterTagText}>{getTypeLabel(t, type)}</span>
+            </span>
+            <span className={styles.filterTagCount}>{typeCounts[type] ?? 0}</span>
+          </button>
+        );
+      })}
+    </div>
+  );
+
+  const titleNode = (
+    <div className={styles.titleWrapper}>
+      <span>{t('auth_files.title_section')}</span>
+      {files.length > 0 && <span className={styles.countBadge}>{files.length}</span>}
+    </div>
+  );
+
+  const deleteAllButtonLabel = (() => {
+    if (disabledOnly || healthyOnly) {
+      return t('auth_files.delete_filtered_result_button');
+    }
+    if (problemOnly) {
+      return normalizedFilter === 'all'
+        ? t('auth_files.delete_problem_button')
+        : t('auth_files.delete_problem_button_with_type', {
+            type: getTypeLabel(t, normalizedFilter),
+          });
+    }
+    return normalizedFilter === 'all'
+      ? t('auth_files.delete_all_button')
+      : `${t('common.delete')} ${getTypeLabel(t, normalizedFilter)}`;
+  })();
+
+  return (
+    <div className={styles.container}>
+      <section className={styles.authFilesShell}>
+        <div className={styles.authFilesHeader}>
+          <div className={styles.authFilesTitle}>{titleNode}</div>
+          <div className={styles.headerActions}>
+            <Button variant="secondary" size="sm" onClick={handleHeaderRefresh} disabled={loading}>
+              {t('common.refresh')}
+            </Button>
+            <Button
+              variant="secondary"
+              size="sm"
+              onClick={() => setAuthJsonPasteOpen(true)}
+              disabled={disableControls || authJsonPasteSaving}
+              loading={authJsonPasteSaving}
+            >
+              {t('auth_files.paste_button')}
+            </Button>
+            <Button
+              size="sm"
+              onClick={handleUploadClick}
+              disabled={disableControls || uploading}
+              loading={uploading}
+            >
+              {t('auth_files.upload_button')}
+            </Button>
+            <Button
+              variant="danger"
+              size="sm"
+              onClick={() =>
+                handleDeleteAll({
+                  filter: normalizedFilter,
+                  problemOnly,
+                  disabledOnly,
+                  healthyOnly,
+                  onResetFilterToAll: () => setFilter('all'),
+                  onResetProblemOnly: () => setProblemOnly(false),
+                  onResetDisabledOnly: () => setDisabledOnly(false),
+                  onResetHealthyOnly: () => setHealthyOnly(false),
+                })
+              }
+              disabled={disableControls || loading || deletingAll}
+              loading={deletingAll}
+            >
+              {deleteAllButtonLabel}
+            </Button>
+            <input
+              ref={fileInputRef}
+              type="file"
+              accept=".json,application/json"
+              multiple
+              style={{ display: 'none' }}
+              onChange={handleFileChange}
+            />
+          </div>
+        </div>
+
+        {error && <div className={styles.errorBox}>{error}</div>}
+
+        <div className={styles.filterSection}>
+          <div className={styles.filterPanel}>
+            <div className={styles.filterPanelTags}>{renderFilterTags()}</div>
+            <div className={styles.filterControlsPanel}>
+              <div className={styles.filterControls}>
+                <div className={styles.filterItem}>
+                  <label>{t('auth_files.search_label')}</label>
+                  <Input
+                    value={search}
+                    onChange={(e) => {
+                      setSearch(e.target.value);
+                      setPage(1);
+                    }}
+                    placeholder={t('auth_files.search_placeholder')}
+                    rightElement={<IconSearch size={16} />}
+                    aria-label={t('auth_files.search_label')}
+                  />
+                </div>
+                <div className={styles.filterItem}>
+                  <label>{t('auth_files.page_size_label')}</label>
+                  <input
+                    className={styles.pageSizeSelect}
+                    type="number"
+                    min={MIN_CARD_PAGE_SIZE}
+                    max={MAX_CARD_PAGE_SIZE}
+                    step={1}
+                    value={pageSizeInput}
+                    onChange={handlePageSizeChange}
+                    onBlur={(e) => commitPageSizeInput(e.currentTarget.value)}
+                    onKeyDown={(e) => {
+                      if (e.key === 'Enter') {
+                        e.currentTarget.blur();
+                      }
+                    }}
+                  />
+                </div>
+                <div className={styles.filterItem}>
+                  <label>{t('auth_files.sort_label')}</label>
+                  <Select
+                    className={styles.sortSelect}
+                    value={sortMode}
+                    options={sortOptions}
+                    onChange={handleSortModeChange}
+                    ariaLabel={t('auth_files.sort_label')}
+                    fullWidth
+                  />
+                </div>
+                <div className={`${styles.filterItem} ${styles.filterToggleItem}`}>
+                  <label>{t('auth_files.display_options_label')}</label>
+                  <div className={styles.filterToggleGroup}>
+                    <div className={styles.filterToggleCard}>
+                      <ToggleSwitch
+                        checked={problemOnly}
+                        onChange={(value) => {
+                          setProblemOnly(value);
+                          if (value) setHealthyOnly(false);
+                          setPage(1);
+                        }}
+                        ariaLabel={t('auth_files.problem_filter_only')}
+                        label={
+                          <span className={styles.filterToggleLabel}>
+                            {t('auth_files.problem_filter_only')}
+                          </span>
+                        }
+                      />
+                    </div>
+                    <div className={styles.filterToggleCard}>
+                      <ToggleSwitch
+                        checked={disabledOnly}
+                        onChange={(value) => {
+                          setDisabledOnly(value);
+                          if (value) setHealthyOnly(false);
+                          setPage(1);
+                        }}
+                        ariaLabel={t('auth_files.disabled_filter_only')}
+                        label={
+                          <span className={styles.filterToggleLabel}>
+                            {t('auth_files.disabled_filter_only')}
+                          </span>
+                        }
+                      />
+                    </div>
+                    <div className={styles.filterToggleCard}>
+                      <ToggleSwitch
+                        checked={healthyOnly}
+                        onChange={(value) => {
+                          setHealthyOnly(value);
+                          if (value) {
+                            setProblemOnly(false);
+                            setDisabledOnly(false);
+                          }
+                          setPage(1);
+                        }}
+                        ariaLabel={t('auth_files.healthy_filter_only')}
+                        label={
+                          <span className={styles.filterToggleLabel}>
+                            {t('auth_files.healthy_filter_only')}
+                          </span>
+                        }
+                      />
+                    </div>
+                    <div className={styles.filterToggleCard}>
+                      <ToggleSwitch
+                        checked={compactMode}
+                        onChange={(value) => setCompactMode(value)}
+                        ariaLabel={t('auth_files.compact_mode_label')}
+                        label={
+                          <span className={styles.filterToggleLabel}>
+                            {t('auth_files.compact_mode_label')}
+                          </span>
+                        }
+                      />
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+
+          <div className={styles.filterContent}>
+            {loading ? (
+              <div className={styles.hint}>{t('common.loading')}</div>
+            ) : pageItems.length === 0 ? (
+              <EmptyState
+                title={t('auth_files.search_empty_title')}
+                description={t('auth_files.search_empty_desc')}
+              />
+            ) : (
+              <div
+                className={`${styles.fileGrid} ${quotaFilterType ? styles.fileGridQuotaManaged : ''} ${compactMode ? styles.fileGridCompact : ''}`}
+              >
+                {pageItems.map((file) => (
+                  <AuthFileCard
+                    key={file.name}
+                    file={file}
+                    compact={compactMode}
+                    selected={selectedFiles.has(file.name)}
+                    resolvedTheme={resolvedTheme}
+                    disableControls={disableControls}
+                    deleting={deleting}
+                    statusUpdating={statusUpdating}
+                    quotaFilterType={quotaFilterType}
+                    statusBarCache={statusBarCache}
+                    onShowModels={showModels}
+                    onDownload={handleDownload}
+                    onOpenPrefixProxyEditor={openPrefixProxyEditor}
+                    onDelete={handleDelete}
+                    onToggleStatus={handleStatusToggle}
+                    onToggleSelect={toggleSelect}
+                  />
+                ))}
+              </div>
+            )}
+
+            {!loading && sorted.length > pageSize && (
+              <div className={styles.pagination}>
+                <Button
+                  variant="secondary"
+                  size="sm"
+                  onClick={() => setPage(Math.max(1, currentPage - 1))}
+                  disabled={currentPage <= 1}
+                >
+                  {t('auth_files.pagination_prev')}
+                </Button>
+                <div className={styles.pageInfo}>
+                  {t('auth_files.pagination_info', {
+                    current: currentPage,
+                    total: totalPages,
+                    count: sorted.length,
+                  })}
+                </div>
+                <Button
+                  variant="secondary"
+                  size="sm"
+                  onClick={() => setPage(Math.min(totalPages, currentPage + 1))}
+                  disabled={currentPage >= totalPages}
+                >
+                  {t('auth_files.pagination_next')}
+                </Button>
+              </div>
+            )}
+          </div>
+        </div>
+      </section>
+
+      <OAuthExcludedCard
+        disableControls={disableControls}
+        excludedError={excludedError}
+        excluded={excluded}
+        onAdd={() => openExcludedEditor()}
+        onEdit={openExcludedEditor}
+        onDelete={deleteExcluded}
+      />
+
+      <OAuthModelAliasCard
+        disableControls={disableControls}
+        viewMode={viewMode}
+        onViewModeChange={setViewMode}
+        onAdd={() => openModelAliasEditor()}
+        onEditProvider={openModelAliasEditor}
+        onDeleteProvider={deleteModelAlias}
+        modelAliasError={modelAliasError}
+        modelAlias={modelAlias}
+        allProviderModels={allProviderModels}
+        onUpdate={handleMappingUpdate}
+        onDeleteLink={handleDeleteLink}
+        onToggleFork={handleToggleFork}
+        onRenameAlias={handleRenameAlias}
+        onDeleteAlias={handleDeleteAlias}
+      />
+
+      <AuthFileModelsModal
+        open={modelsModalOpen}
+        fileName={modelsFileName}
+        fileType={modelsFileType}
+        loading={modelsLoading}
+        error={modelsError}
+        models={modelsList}
+        excluded={excluded}
+        onClose={closeModelsModal}
+        onCopyText={copyTextWithNotification}
+      />
+
+      <AuthFilesPrefixProxyEditorModal
+        disableControls={disableControls}
+        editor={prefixProxyEditor}
+        updatedText={prefixProxyUpdatedText}
+        dirty={prefixProxyDirty}
+        onClose={closePrefixProxyEditor}
+        onCopyText={copyTextWithNotification}
+        onSave={handlePrefixProxySave}
+        onChange={handlePrefixProxyChange}
+      />
+
+      <AuthJsonPasteModal
+        open={authJsonPasteOpen}
+        saving={authJsonPasteSaving}
+        disabled={disableControls}
+        onClose={() => {
+          if (!authJsonPasteSaving) setAuthJsonPasteOpen(false);
+        }}
+        onSave={handleSavePastedAuthJson}
+      />
+
+      {batchActionBarVisible && typeof document !== 'undefined'
+        ? createPortal(
+            <div className={styles.batchActionContainer} ref={floatingBatchActionsRef}>
+              <div className={styles.batchActionBar}>
+                <div className={styles.batchActionLeft}>
+                  <span className={styles.batchSelectionText}>
+                    {t('auth_files.batch_selected', { count: selectionCount })}
+                  </span>
+                  <Button
+                    variant="secondary"
+                    size="sm"
+                    onClick={() => selectAllVisible(pageItems)}
+                    disabled={selectablePageItems.length === 0}
+                  >
+                    {t('auth_files.batch_select_page')}
+                  </Button>
+                  <Button
+                    variant="secondary"
+                    size="sm"
+                    onClick={() => selectAllVisible(sorted)}
+                    disabled={selectableFilteredItems.length === 0}
+                  >
+                    {t('auth_files.batch_select_filtered')}
+                  </Button>
+                  <Button
+                    variant="ghost"
+                    size="sm"
+                    onClick={() => invertVisibleSelection(pageItems)}
+                    disabled={selectablePageItems.length === 0}
+                  >
+                    {t('auth_files.batch_invert_page')}
+                  </Button>
+                  <Button variant="ghost" size="sm" onClick={deselectAll}>
+                    {t('auth_files.batch_deselect')}
+                  </Button>
+                </div>
+                <div className={styles.batchActionRight}>
+                  <Button
+                    variant="secondary"
+                    size="sm"
+                    onClick={() => void batchDownload(selectedNames)}
+                    disabled={disableControls || selectedNames.length === 0}
+                  >
+                    {t('auth_files.batch_download')}
+                  </Button>
+                  <Button
+                    size="sm"
+                    onClick={() => batchSetStatus(selectedNames, true)}
+                    disabled={batchStatusButtonsDisabled}
+                  >
+                    {t('auth_files.batch_enable')}
+                  </Button>
+                  <Button
+                    variant="secondary"
+                    size="sm"
+                    onClick={() => batchSetStatus(selectedNames, false)}
+                    disabled={batchStatusButtonsDisabled}
+                  >
+                    {t('auth_files.batch_disable')}
+                  </Button>
+                  <Button
+                    variant="danger"
+                    size="sm"
+                    onClick={() => batchDelete(selectedNames)}
+                    disabled={disableControls || selectedNames.length === 0}
+                  >
+                    {t('common.delete')}
+                  </Button>
+                </div>
+              </div>
+            </div>,
+            document.body
+          )
+        : null}
+    </div>
+  );
+}
