+++ src/features/authFiles/components/AuthFileModelsModal.tsx
@@ -4,7 +4,7 @@ import { Button } from '@/components/ui/Button';
 import { EmptyState } from '@/components/ui/EmptyState';
 import type { AuthFileModelItem } from '@/features/authFiles/constants';
 import { isModelExcluded } from '@/features/authFiles/constants';
-import styles from '@/pages/AuthFilesPage.module.scss';
+import styles from '@/features/authFiles/AuthFilesPage.module.scss';
 
 export type AuthFileModelsModalProps = {
   open: boolean;
@@ -88,4 +88,3 @@ export function AuthFileModelsModal(props: AuthFileModelsModalProps) {
     </Modal>
   );
 }
-
