+++ src/features/authFiles/components/OAuthModelAliasCard.tsx
@@ -7,7 +7,7 @@ import { ModelMappingDiagram, type ModelMappingDiagramRef } from '@/components/m
 import { IconChevronUp } from '@/components/ui/icons';
 import type { OAuthModelAliasEntry } from '@/types';
 import type { AuthFileModelItem } from '@/features/authFiles/constants';
-import styles from '@/pages/AuthFilesPage.module.scss';
+import styles from '@/features/authFiles/AuthFilesPage.module.scss';
 
 type UnsupportedError = 'unsupported' | null;
 type ViewMode = 'diagram' | 'list';
@@ -149,4 +149,3 @@ export function OAuthModelAliasCard(props: OAuthModelAliasCardProps) {
     </Card>
   );
 }
-
